feat: NetworkCheck 연결상태 모니터링 기능 구현
